+++ elastic-job-spring/src/main/java/com/dangdang/ddframe/job/spring/namespace/parser/common/AbstractJobBeanDefinitionParser.java
@@ -33,6 +33,25 @@ import org.w3c.dom.Element;
 
 import java.util.List;
 
+import static com.dangdang.ddframe.job.spring.namespace.constants.BaseJobBeanDefinitionParserTag.COMPLETED_TIMEOUT_MILLISECONDS_ATTRIBUTE;
+import static com.dangdang.ddframe.job.spring.namespace.constants.BaseJobBeanDefinitionParserTag.CRON_ATTRIBUTE;
+import static com.dangdang.ddframe.job.spring.namespace.constants.BaseJobBeanDefinitionParserTag.CLASS_ATTRIBUTE;
+import static com.dangdang.ddframe.job.spring.namespace.constants.BaseJobBeanDefinitionParserTag.DESCRIPTION_ATTRIBUTE;
+import static com.dangdang.ddframe.job.spring.namespace.constants.BaseJobBeanDefinitionParserTag.DISABLED_ATTRIBUTE;
+import static com.dangdang.ddframe.job.spring.namespace.constants.BaseJobBeanDefinitionParserTag.FAILOVER_ATTRIBUTE;
+import static com.dangdang.ddframe.job.spring.namespace.constants.BaseJobBeanDefinitionParserTag.JOB_PARAMETER_ATTRIBUTE;
+import static com.dangdang.ddframe.job.spring.namespace.constants.BaseJobBeanDefinitionParserTag.JOB_SHARDING_STRATEGY_CLASS_ATTRIBUTE;
+import static com.dangdang.ddframe.job.spring.namespace.constants.BaseJobBeanDefinitionParserTag.LISTENER_TAG;
+import static com.dangdang.ddframe.job.spring.namespace.constants.BaseJobBeanDefinitionParserTag.MAX_TIME_DIFF_SECONDS_ATTRIBUTE;
+import static com.dangdang.ddframe.job.spring.namespace.constants.BaseJobBeanDefinitionParserTag.MISFIRE_ATTRIBUTE;
+import static com.dangdang.ddframe.job.spring.namespace.constants.BaseJobBeanDefinitionParserTag.MONITOR_EXECUTION_ATTRIBUTE;
+import static com.dangdang.ddframe.job.spring.namespace.constants.BaseJobBeanDefinitionParserTag.MONITOR_PORT_ATTRIBUTE;
+import static com.dangdang.ddframe.job.spring.namespace.constants.BaseJobBeanDefinitionParserTag.OVERWRITE_ATTRIBUTE;
+import static com.dangdang.ddframe.job.spring.namespace.constants.BaseJobBeanDefinitionParserTag.SHARDING_ITEM_PARAMETERS_ATTRIBUTE;
+import static com.dangdang.ddframe.job.spring.namespace.constants.BaseJobBeanDefinitionParserTag.SHARDING_TOTAL_COUNT_ATTRIBUTE;
+import static com.dangdang.ddframe.job.spring.namespace.constants.BaseJobBeanDefinitionParserTag.STARTED_TIMEOUT_MILLISECONDS_ATTRIBUTE;
+import static com.dangdang.ddframe.job.spring.namespace.constants.ScriptJobBeanDefinitionParserTag.SCRIPT_COMMAND_LINE_ATTRIBUTE;
+
 /**
  * 基本作业的命名空间解析器.
  * 
@@ -45,7 +64,7 @@ public abstract class AbstractJobBeanDefinitionParser extends AbstractBeanDefini
     protected AbstractBeanDefinition parseInternal(final Element element, final ParserContext parserContext) {
         BeanDefinitionBuilder factory = BeanDefinitionBuilder.rootBeanDefinition(SpringJobScheduler.class);
         factory.setInitMethodName("init");
-        factory.addConstructorArgReference(element.getAttribute("regCenter"));
+        factory.addConstructorArgReference(element.getAttribute("reg-center"));
         factory.addConstructorArgReference(createJobConfiguration(element, parserContext));
         factory.addConstructorArgValue(createJobListeners(element));
         return factory.getBeanDefinition();
@@ -53,29 +72,28 @@ public abstract class AbstractJobBeanDefinitionParser extends AbstractBeanDefini
     
     private String createJobConfiguration(final Element element, final ParserContext parserContext) {
         BeanDefinitionBuilder factory = BeanDefinitionBuilder.rootBeanDefinition(getJobConfigurationDTO());
-        factory.addConstructorArgValue(element.getAttribute("id"));
+        factory.addConstructorArgValue(element.getAttribute(ID_ATTRIBUTE));
         if (!getJobConfigurationDTO().isAssignableFrom(ScriptJobConfigurationDto.class)) {
-            factory.addConstructorArgValue(element.getAttribute("class"));
+            factory.addConstructorArgValue(element.getAttribute(CLASS_ATTRIBUTE));
         }
-        factory.addConstructorArgValue(element.getAttribute("shardingTotalCount"));
-        factory.addConstructorArgValue(element.getAttribute("cron"));
+        factory.addConstructorArgValue(element.getAttribute(SHARDING_TOTAL_COUNT_ATTRIBUTE));
+        factory.addConstructorArgValue(element.getAttribute(CRON_ATTRIBUTE));
         if (getJobConfigurationDTO().isAssignableFrom(ScriptJobConfigurationDto.class)) {
-            factory.addConstructorArgValue(element.getAttribute("cron"));    
+            factory.addConstructorArgValue(element.getAttribute(SCRIPT_COMMAND_LINE_ATTRIBUTE));
         }
-        addPropertyValueIfNotEmpty("shardingItemParameters", element, factory);
-        addPropertyValueIfNotEmpty("jobParameter", element, factory);
-        addPropertyValueIfNotEmpty("monitorExecution", element, factory);
-        addPropertyValueIfNotEmpty("monitorPort", element, factory);
-        addPropertyValueIfNotEmpty("fetchDataCount", element, factory);
-        addPropertyValueIfNotEmpty("maxTimeDiffSeconds", element, factory);
-        addPropertyValueIfNotEmpty("failover", element, factory);
-        addPropertyValueIfNotEmpty("misfire", element, factory);
-        addPropertyValueIfNotEmpty("jobShardingStrategyClass", element, factory);
-        addPropertyValueIfNotEmpty("description", element, factory);
-        addPropertyValueIfNotEmpty("disabled", element, factory);
-        addPropertyValueIfNotEmpty("overwrite", element, factory);
+        addPropertyValueIfNotEmpty(SHARDING_ITEM_PARAMETERS_ATTRIBUTE, "shardingItemParameters", element, factory);
+        addPropertyValueIfNotEmpty(JOB_PARAMETER_ATTRIBUTE, "jobParameter", element, factory);
+        addPropertyValueIfNotEmpty(MONITOR_EXECUTION_ATTRIBUTE, "monitorExecution", element, factory);
+        addPropertyValueIfNotEmpty(MONITOR_PORT_ATTRIBUTE, "monitorPort", element, factory);
+        addPropertyValueIfNotEmpty(MAX_TIME_DIFF_SECONDS_ATTRIBUTE, "maxTimeDiffSeconds", element, factory);
+        addPropertyValueIfNotEmpty(FAILOVER_ATTRIBUTE, "failover", element, factory);
+        addPropertyValueIfNotEmpty(MISFIRE_ATTRIBUTE, "misfire", element, factory);
+        addPropertyValueIfNotEmpty(JOB_SHARDING_STRATEGY_CLASS_ATTRIBUTE, "jobShardingStrategyClass", element, factory);
+        addPropertyValueIfNotEmpty(DESCRIPTION_ATTRIBUTE, "description", element, factory);
+        addPropertyValueIfNotEmpty(DISABLED_ATTRIBUTE, "disabled", element, factory);
+        addPropertyValueIfNotEmpty(OVERWRITE_ATTRIBUTE, "overwrite", element, factory);
         setPropertiesValue(element, factory);
-        String result = element.getAttribute("id") + "Conf";
+        String result = element.getAttribute(ID_ATTRIBUTE) + "Conf";
         parserContext.getRegistry().registerBeanDefinition(result, factory.getBeanDefinition());
         return result;
     }
@@ -85,17 +103,17 @@ public abstract class AbstractJobBeanDefinitionParser extends AbstractBeanDefini
     protected abstract void setPropertiesValue(final Element element, final BeanDefinitionBuilder factory);
     
     private List<BeanDefinition> createJobListeners(final Element element) {
-        List<Element> listenerElements = DomUtils.getChildElementsByTagName(element, "listener");
+        List<Element> listenerElements = DomUtils.getChildElementsByTagName(element, LISTENER_TAG);
         List<BeanDefinition> result = new ManagedList<>(listenerElements.size());
         for (Element each : listenerElements) {
-            String className = each.getAttribute("class");
+            String className = each.getAttribute(CLASS_ATTRIBUTE);
             BeanDefinitionBuilder factory = BeanDefinitionBuilder.rootBeanDefinition(className);
             factory.setScope(BeanDefinition.SCOPE_PROTOTYPE);
             try {
                 Class listenerClass = Class.forName(className);
                 if (AbstractDistributeOnceElasticJobListener.class.isAssignableFrom(listenerClass)) {
-                    factory.addConstructorArgValue(each.getAttribute("startedTimeoutMilliseconds"));
-                    factory.addConstructorArgValue(each.getAttribute("completedTimeoutMilliseconds"));
+                    factory.addConstructorArgValue(each.getAttribute(STARTED_TIMEOUT_MILLISECONDS_ATTRIBUTE));
+                    factory.addConstructorArgValue(each.getAttribute(COMPLETED_TIMEOUT_MILLISECONDS_ATTRIBUTE));
                 }
             } catch (final ClassNotFoundException ex) {
                 throw new RuntimeException(ex);
@@ -105,10 +123,10 @@ public abstract class AbstractJobBeanDefinitionParser extends AbstractBeanDefini
         return result;
     }
     
-    protected void addPropertyValueIfNotEmpty(final String propertyName, final Element element, final BeanDefinitionBuilder factory) {
-        String propertyValue = element.getAttribute(propertyName);
-        if (!Strings.isNullOrEmpty(propertyValue)) {
-            factory.addPropertyValue(propertyName, propertyValue);
+    protected final void addPropertyValueIfNotEmpty(final String attributeName, final String propertyName, final Element element, final BeanDefinitionBuilder factory) {
+        String attributeValue = element.getAttribute(attributeName);
+        if (!Strings.isNullOrEmpty(attributeValue)) {
+            factory.addPropertyValue(propertyName, attributeValue);
         }
     }